descrição abaixo;
- obrigatoriedade da seleção de serviços
- acerto no cadastro de lote com imagem
- possibilidade de poder editar valor do campo redirecionamento
- não exibir o valor no email quando tarifa por serviços
- adição do campo valor item voi s no produto do lote
- adição do campo rank no produto do lote
- adição do campo, calculado, rank médio no lote
- valor default da tarifa amazon igual a zero
- adição da informação de lucro na listagem de lotes

+++ utils/storage.py
@@ -11,5 +11,6 @@ class OverWriteStorage(FileSystemStorage):
     def get_available_name(self, name, max_length=None):
         logger.debug('@@@@@@@@@@ STORAGE @@@@@@@@@@@@@')
         dir_name, file_name = os.path.split(name)
-        shutil.rmtree(os.path.join(settings.MEDIA_ROOT, dir_name))
+        if os.path.exists(dir_name):
+            shutil.rmtree(os.path.join(settings.MEDIA_ROOT, dir_name))
         return name
